requirements.txt
streamlit>=1.33.0
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.36
ta>=0.11.0
requests>=2.31.0
pyarrow>=15.0.0
openai>=1.30.0
matplotlib>=3.8.0,<3.9

python-dotenv>=1.0.1
